+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.43.33](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.32...@synapsecns/synapse-interface@0.43.33) (2025-11-27)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.43.32](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.31...@synapsecns/synapse-interface@0.43.32) (2025-11-27)